+++ embassy-executor/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Replaced Pender. Implementations now must define an extern function called `__pender`.
+- Added `SpawnToken::is_success()`
 
 ## 0.2.1 - 2023-08-10